Address comments by @ikhoon
