+++ spring/boot-webflux-autoconfigure/build.gradle
@@ -28,7 +28,7 @@ dependencies {
 task copyFiles(type: Copy) {
     from "${rootProject.projectDir}/spring/boot-autoconfigure/src/main/java"
     into "${project.ext.genSrcDir}/main/java"
-    include '**/ArmeriaBeanPostProcessor.java'
+    include '**/*ArmeriaBeanPostProcessor.java'
     include '**/*ExampleHeaders.java'
     include '**/*ExampleRequest.java'
     include '**/*RegistrationBean.java'
